add comparing ID in all equals() method(+hashCode())

+++ src/main/java/Room.java
@@ -34,6 +34,6 @@ public class Room {
 
     @Override
     public int hashCode() {
-        return number;
+        return number.intValue();
     }
 }
